requirements.txt
torch
torchvision
torchaudio
numpy
pyyaml
gymnasium
dm-control
opencv-python